cloud_functions/requirements.txt
# Function dependencies, for example:
# package>=version
firebase_admin
redis
google-cloud-pubsub
